requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pandas==2.2.0
numpy==1.26.4
python-dateutil==2.8.2
pytz==2023.3
websocket-client==1.6.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
